move hidden spaces back to being filtered in the UI
